+++ src/Statistiques/register.ts
